SDC-3114. Fix max cache size to be unlimited if it is -1.
Reviewed-on: https://review.streamsets.net/3080
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>

+++ hive-protolib/src/main/java/com/streamsets/pipeline/stage/destination/hive/HiveConfigBean.java
@@ -82,5 +82,5 @@ public class HiveConfigBean {
       displayPosition = 30,
       group = "ADVANCED"
   )
-  public int maxCacheSize = -1;
+  public long maxCacheSize = -1L;
 }
